requirements.txt
flask==2.3.3
cohere==5.3.4
python-dotenv==1.0.0
